aats/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
aiohttp>=3.8.5
websockets>=11.0.3
asyncio>=3.4.3
tenacity>=8.2.2

# Database Adapters
psycopg2-binary>=2.9.7
pymongo>=4.4.1
neo4j>=5.11.0
redis>=5.0.0
motor>=3.3.1

# Message Brokers
kafka-python>=2.0.2
pika>=1.3.1

# Search Engines
elasticsearch>=8.9.0
pymilvus>=2.3.0

# API Gateway & Service Mesh
requests>=2.31.0
istio-api>=1.19.0
kubernetes>=27.2.0

# Monitoring & Observability
prometheus-client>=0.17.1
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
grafana-api>=1.0.3
python-alertmanager>=1.0.0

# Machine Learning & AI
openai>=1.0.0
anthropic>=0.3.0
cohere>=4.21
transformers>=4.33.2
torch>=2.0.1
tensorflow>=2.13.0
langchain>=0.0.300
chromadb>=0.4.6

# Integration Services
atlassian-python-api>=3.41.1
PyGithub>=2.1.1
slack-sdk>=3.22.0

# Security
cryptography>=41.0.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Documentation
sphinx>=7.1.2
mkdocs>=1.5.2
mkdocs-material>=9.2.7

# Development Tools
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0
pre-commit>=3.3.3

# Utilities
PyYAML>=6.0.1
python-multipart>=0.0.6
python-json-logger>=2.0.7
structlog>=23.1.0
rich>=13.5.2
typer>=0.9.0
